fix: create test output folder if it doesn't exist already
if the folder didn't exist, the step for creating a db_dump will fail.

+++ test/test_build.bat
@@ -53,7 +53,8 @@ call setup_sqldatamart.bat
 @SET pg_database_old=%PGDATABASE%
 @SET PGDATABASE=openclinica_fdw_db
 
-del /F /Q %export_file_path%
+if not exist %test_output_path% mkdir %test_output_path%
+if exist %export_file_path% del /F /Q %export_file_path%
 
 psql -f %~dp0util_dm_copy_schema_to_tables_for_export.sql
 psql -c "SELECT %export_func%($s$%target_schema_name%$s$);" -P pager
